Refactoring: add a Handler class, so it can be reused across different concurrency strategies
