+++ lib/adelnor/base_server.rb
@@ -6,6 +6,7 @@ require 'stringio'
 
 require_relative './request'
 require_relative './response'
+require_relative './handler'
 
 module Adelnor
   class BaseServer
@@ -19,6 +20,8 @@ module Adelnor
       @socket.bind(addr)
       @socket.listen(2)
 
+      @handler = Handler.new(@rack_app, @port)
+
       puts welcome_message
     end
 
@@ -26,54 +29,13 @@ module Adelnor
       new(*args).run
     end
 
-    def rack_data(request)
-      {
-        'REQUEST_METHOD' => request.request_method,
-        'PATH_INFO' => request.path_info,
-        'QUERY_STRING' => request.query_string,
-        'SERVER_PORT' => @port,
-        'SERVER_NAME' => request.headers['Host'],
-        'CONTENT_LENGTH' => request.content_length,
-        'HTTP_COOKIE' => request.headers['Cookie'],
-        'rack.input' => StringIO.new(request.body)
-      }.merge(request.headers)
-    end
-
     def run
       loop do
         client, = @socket.accept
 
-        handle(client)
+        @handler.handle(client)
         client.close
       end
-    end
-
-    def handle(client)
-      read_request_message(client)
-        .then { |message|  Request.build(message) }
-        .tap  { |request|  request.parse_body!(client) }
-        .then { |request|  rack_data(request) }
-        .then { |data|     @rack_app.call(data) }
-        .then { |result|   Response.build(*result) }
-        .then { |response| client.puts(response) }
-    end
-
-    def read_request_message(client)
-      message = ''
-
-      if (line = client.gets)
-        message += line
-      end
-
-      puts "\n[#{Time.now}] #{message}"
-
-      while (line = client.gets)
-        break if line == "\r\n"
-
-        message += line
-      end
-
-      message
     end
 
     def welcome_message
