+++ lib/adelnor/clustered_server.rb
@@ -4,14 +4,8 @@ require_relative './base_server'
 
 module Adelnor
   class ClusteredServer < BaseServer
-    def initialize(rack_app, port, options = {})
-      super(rack_app, port, options)
-
-      @workers = options[:workers]
-    end
-
     def run
-      @workers.times do
+      @options[:workers].times do
         fork { handle_worker }
       end
 
@@ -25,7 +19,7 @@ module Adelnor
       loop do
         client, = @socket.accept
 
-        handle(client)
+        @handler.handle(client)
         client.close
       end
     end
